refactor(api): move sco-organization-invitation serializer
to src
