+++ api/tests/team/unit/infrastructure/serializers/jsonapi/sco-organization-invitation-serializer_test.js
@@ -1,7 +1,7 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/sco-organization-invitation-serializer.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../src/team/infrastructure/serializers/jsonapi/sco-organization-invitation.serializer.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
-describe('Unit | Serializer | JSONAPI | sco-organization-invitation-serializer', function () {
+describe('Unit | Team | Infrastructure | Serializer | JSONAPI | sco-organization-invitation', function () {
   describe('#serialize', function () {
     // TODO: Fix this the next time the file is edited.
     // eslint-disable-next-line mocha/no-setup-in-describe
